+++ packages/sdks-tests/src/e2e-tests/editing.spec.ts
@@ -22,7 +22,11 @@ import { CUSTOM_COMPONENT_NO_DEFAULT_VALUE } from '../specs/custom-component-no-
 import { SYMBOLS_WITH_LIST_CONTENT_INPUT } from '../specs/symbols-with-list-content-input.js';
 import { NEW_BLOCK_ADD, NEW_BLOCK_ADD_2 } from '../specs/new-block-add.js';
 import { SECTION_CHILDREN } from '../specs/section-children.js';
-
+import {
+  GET_CONTENT_SYMBOL_UPDATE_ENTRY_ONE,
+  GET_CONTENT_SYMBOL_UPDATE_ENTRY__TWO,
+  MAIN_CONTENT,
+} from '../specs/get-content-symbol-update-entry.js';
 const editorTests = ({
   noTrustedHosts,
   editorIsInViewPort,
@@ -549,6 +553,10 @@ test.describe('Visual Editing', () => {
       test.skip(checkIsGen1React(sdk));
       test.skip(packageName === 'nextjs-sdk-next-app');
       test.skip(
+        packageName === 'qwik-city' || packageName === 'nuxt',
+        'Failing on the CI: Test timeout of 30000ms exceeded'
+      );
+      test.skip(
         packageName === 'vue',
         `Failing on the CI: TypeError: Cannot read properties of null (reading 'namespaceURI')`
       );
@@ -900,4 +908,43 @@ test.describe('Visual Editing', () => {
       await page.frameLocator('iframe').getByText('new text').waitFor({ state: 'hidden' });
     });
   });
+});
+
+test('Symbol should update the data when entry is updated', async ({
+  page,
+  sdk,
+  packageName,
+  basePort,
+}) => {
+  test.skip(excludeGen1(sdk));
+  test.skip(
+    packageName === 'nextjs-sdk-next-app' || packageName === 'remix',
+    'This both packages are SSR hence getting symbol not found error'
+  );
+  test.skip(packageName === 'qwik-city', 'Qwik-city API not getting called');
+  const urlMatch = /https:\/\/cdn\.builder\.io\/api\/v3\/content\/symbol/;
+  await page.route(urlMatch, route => {
+    const url = new URL(route.request().url());
+    const queryIdParam = url.searchParams.get('query.id');
+    let symbolToReturn = null;
+    if (queryIdParam === '"e2a166f7d9544ed9ade283abf9491af3"') {
+      symbolToReturn = GET_CONTENT_SYMBOL_UPDATE_ENTRY_ONE;
+    } else if (queryIdParam === '"aa024e6851e94b49b99f41a2294fd423"') {
+      symbolToReturn = GET_CONTENT_SYMBOL_UPDATE_ENTRY__TWO;
+    } else {
+      throw new Error(`Unknown query id: ${queryIdParam}`);
+    }
+    return route.fulfill({
+      status: 200,
+      json: {
+        results: [symbolToReturn],
+      },
+    });
+  });
+  await launchEmbedderAndWaitForSdk({ path: '/symbol-update-entries', basePort, page, sdk });
+  const newContent = cloneContent(MAIN_CONTENT);
+  await page.frameLocator('iframe').getByText('Green Potato').waitFor();
+  newContent.data.blocks[0].component.options.symbol.entry = 'aa024e6851e94b49b99f41a2294fd423';
+  await sendContentUpdateMessage({ page, newContent, model: 'page' });
+  await page.frameLocator('iframe').getByText('Red tomato').waitFor();
 });