+++ packages/sdks/src/blocks/symbol/symbol.lite.tsx
@@ -73,8 +73,10 @@ export default function Symbol(props: SymbolProps) {
           builderContextValue: props.builderContext.value,
         }))) as Nullable<BuilderContent>,
     }),
+    symbolEntry: props.symbol?.entry,
     setContent() {
-      if (state.contentToUse) return;
+      if (state.contentToUse && state.symbolEntry === props.symbol?.entry)
+        return;
 
       fetchSymbolContent({
         symbol: props.symbol,
@@ -82,6 +84,7 @@ export default function Symbol(props: SymbolProps) {
       }).then((newContent) => {
         if (newContent) {
           state.contentToUse = newContent;
+          state.symbolEntry = props.symbol?.entry;
         }
       });
     },